pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netbox-graphql-query-optimizer"
version = "0.1.0"
description = "Static analysis tool for NetBox GraphQL queries"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "NetBox GraphQL Query Optimizer Contributors"}
]
keywords = ["netbox", "graphql", "performance", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "graphql-core>=3.2.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
netbox-gqo = "netbox_graphql_query_optimizer.cli:main"

[tool.setuptools]
packages = ["netbox_graphql_query_optimizer"]

[tool.black]
line-length = 100
target-version = ['py310']
